refactor: 공용으로 사용 가능한 클래스 common 패키지로 이동
