+++ src/main/java/com/commerceteamproject/admin/enitity/Admin.java
@@ -1,6 +1,6 @@
 package com.commerceteamproject.admin.enitity;
 
-import com.commerceteamproject.common.BaseEntity;
+import com.commerceteamproject.common.entity.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
