+++ src/main/java/com/commerceteamproject/admin/exception/DuplicateEmailException.java
@@ -1,6 +1,6 @@
 package com.commerceteamproject.admin.exception;
 
-import com.commerceteamproject.common.ServiceException;
+import com.commerceteamproject.common.exception.ServiceException;
 import org.springframework.http.HttpStatus;
 
 public class DuplicateEmailException extends ServiceException {
