+++ src/main/java/com/commerceteamproject/admin/exception/EmailNotFoundException.java
@@ -1,6 +1,6 @@
 package com.commerceteamproject.admin.exception;
 
-import com.commerceteamproject.common.ServiceException;
+import com.commerceteamproject.common.exception.ServiceException;
 import org.springframework.http.HttpStatus;
 
 public class EmailNotFoundException extends ServiceException {
