+++ src/main/java/com/commerceteamproject/common/exception/AccessDeniedException.java
@@ -1,6 +1,5 @@
-package com.commerceteamproject.customer.exception;
+package com.commerceteamproject.common.exception;
 
-import com.commerceteamproject.common.ServiceException;
 import org.springframework.http.HttpStatus;
 
 public class AccessDeniedException extends ServiceException {
