+++ src/main/java/com/commerceteamproject/common/exception/ServiceException.java
@@ -1,4 +1,4 @@
-package com.commerceteamproject.common;
+package com.commerceteamproject.common.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
