+++ src/main/java/com/commerceteamproject/config/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.commerceteamproject.config;
 
-import com.commerceteamproject.common.ServiceException;
+import com.commerceteamproject.common.exception.ServiceException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
