+++ src/main/java/com/commerceteamproject/customer/exception/CustomerNotFoundException.java
@@ -1,6 +1,6 @@
 package com.commerceteamproject.customer.exception;
 
-import com.commerceteamproject.common.ServiceException;
+import com.commerceteamproject.common.exception.ServiceException;
 import org.springframework.http.HttpStatus;
 
 public class CustomerNotFoundException extends ServiceException {
